Moved plugin from DQM/SiStripMonitorSummary shared library
The IB quality monitoring found the same plugin in different
plugin shared libraries.
