+++ DQM/TrackingMonitorClient/BuildFile.xml
@@ -11,7 +11,6 @@
 <use   name="DQM/TrackerCommon"/>
 <use   name="DQM/SiStripCommon"/>
 <use   name="DQM/SiStripMonitorClient"/>
-<use   name="DQM/SiStripMonitorSummary"/>
 <use   name="DataFormats/SiStripDetId"/>
 <use   name="boost"/>
 <use   name="root"/>
