+++ DQM/TrackingMonitorClient/plugins/BuildFile.xml
@@ -10,7 +10,6 @@
 <use   name="CalibTracker/Records"/>
 <use   name="DQM/TrackerCommon"/>
 <use   name="DQM/SiStripCommon"/>
-<use   name="DQM/SiStripMonitorSummary"/>
 <use   name="DQM/SiStripMonitorClient"/>
 <use   name="DQM/TrackingMonitorClient"/>
 <use   name="DataFormats/SiStripDetId"/>
